+++ desmost/src/magic/local/colour.ts
@@ -7,6 +7,7 @@ import { Ast } from "../../parser";
 
 type DesmosColourName = keyof typeof Desmos.Colors;
 type DesmosColour = (typeof Desmos.Colors)[DesmosColourName];
+type HexColour = string;
 
 const VALID_COLOURS = ["RED", "BLUE", "GREEN", "PURPLE", "ORANGE", "BLACK"];
 
@@ -17,17 +18,17 @@ export class ColourIncantation extends ArgIncantation<LOCAL>
 	override readonly identifier   = "colour"
 	override readonly alias        = "color"
 	override readonly requires_arg = true
-	override readonly arg_type     = Incantation.ArgType.ENUM
+	override readonly arg_type     = Incantation.ArgType.STRING
 	override readonly description
 		= "Change the colour of a rendered expression, such as a line, region, polygon, etc."
 
-	override apply(target: Desmos.ExpressionState, data: DesmosColour)
+	override apply(target: Desmos.ExpressionState, data: DesmosColour | HexColour)
 	{
 		super.require_expr_type(target.type, "expression");
 		target.color = data;
 	}
 
-	override evaluate_arg(data: string, options: DesmostOptions): DesmosColour
+	override evaluate_arg(data: string, options: DesmostOptions): DesmosColour | HexColour
 	{
 		let colour = data.trim().toUpperCase();
 
@@ -36,15 +37,16 @@ export class ColourIncantation extends ArgIncantation<LOCAL>
 		}
 		else {
 			let internal = Desmos.Colors[colour as DesmosColourName];
+			if (internal != undefined) return internal;
 
-			if (options.check_args && internal == undefined) {
+			if (options.check_args && !colour.startsWith("#")) {
 				throw new DesmostError.InvalidArgument({
 					msg:  `/colour received invalid colour: \`${data}\``,
 					hint: `Valid colours are ${VALID_COLOURS.join(", ")}`,
 				});
 			}
 
-			return internal;
+			return colour;
 		}
 	}