+++ desmost/src/magic/local/dotted.ts
@@ -3,28 +3,28 @@ import { Incantation, type LOCAL } from "../incantation";
 import { Ast } from "../../parser";
 
 
-export class NoLineIncantation extends Incantation<LOCAL>
+export class DottedIncantation extends Incantation<LOCAL>
 {
-	override readonly effect     = Incantation.Effect.LOCAL
-	override readonly identifier = "no-line"
+	override readonly effect       = Incantation.Effect.LOCAL
+	override readonly identifier = "dotted"
 	override readonly description
-		= "Disable rendering lines for a block."
+		= "Render a block as a dotted line."
 
 	override apply(target: Desmos.ExpressionState)
 	{
 		super.require_expr_type(target.type, "expression");
-		target.lines = false;
+		target.lineStyle = Desmos.Styles.DOTTED;
 	}
-		
+	
 	override extract(target: Desmos.Expression): Ast.IncantationInvocation<LOCAL> | void
 	{
-		if (target.lines === false) {
+		if (target.lineStyle === Desmos.Styles.DOTTED) {
 			return {
 				kind: Ast.Kind.INCANTATION_INVOCATION,
-				incantation: no_line,
+				incantation: dotted,
 			};
 		}
 	}
 }
 
-export const no_line = new NoLineIncantation();
+export const dotted = new DottedIncantation();