+++ desmost/src/magic/local/index.ts
@@ -3,10 +3,12 @@ import { Incantation, type LOCAL } from "../incantation";
 import { anim    } from "./anim";
 import { colour  } from "./colour";
 import { dashed  } from "./dashed";
+import { dotted  } from "./dotted";
 import { fill    } from "./fill";
 import { hide    } from "./hide";
 import { label   } from "./label";
 import { line    } from "./line";
+import { no_fill } from "./no-fill";
 import { no_line } from "./no-line";
 import { point   } from "./point";
 import { secret  } from "./secret";
@@ -17,10 +19,12 @@ export const LOCAL_INCANTATIONS = {
 	anim,
 	colour,
 	dashed,
+	dotted,
 	fill,
 	hide,
 	label,
 	line,
+	no_fill,
 	no_line,
 	point,
 	secret,