+++ desmost/src/magic/local/no-fill.ts
@@ -3,28 +3,28 @@ import { Incantation, type LOCAL } from "../incantation";
 import { Ast } from "../../parser";
 
 
-export class NoLineIncantation extends Incantation<LOCAL>
+export class NoFillIncantation extends Incantation<LOCAL>
 {
 	override readonly effect     = Incantation.Effect.LOCAL
-	override readonly identifier = "no-line"
+	override readonly identifier = "no-fill"
 	override readonly description
-		= "Disable rendering lines for a block."
+		= "Disable rendering fill for a block."
 
 	override apply(target: Desmos.ExpressionState)
 	{
 		super.require_expr_type(target.type, "expression");
-		target.lines = false;
+		target.fill = false;
 	}
 		
 	override extract(target: Desmos.Expression): Ast.IncantationInvocation<LOCAL> | void
 	{
-		if (target.lines === false) {
+		if (target.fill === false) {
 			return {
 				kind: Ast.Kind.INCANTATION_INVOCATION,
-				incantation: no_line,
+				incantation: no_fill,
 			};
 		}
 	}
 }
 
-export const no_line = new NoLineIncantation();
+export const no_fill = new NoFillIncantation();